+++ src/main/java/appeng/hooks/DispenserBehaviorTinyTNT.java
@@ -20,8 +20,8 @@ final public class DispenserBehaviorTinyTNT extends BehaviorDefaultDispenseItem
 		int i = dispenser.getXInt() + enumfacing.getFrontOffsetX();
 		int j = dispenser.getYInt() + enumfacing.getFrontOffsetY();
 		int k = dispenser.getZInt() + enumfacing.getFrontOffsetZ();
-		EntityTinyTNTPrimed primedTinyTNTEntity = new EntityTinyTNTPrimed( world, (double) ((float) i + 0.5F), (double) ((float) j + 0.5F),
-				(double) ((float) k + 0.5F), (EntityLiving) null );
+		EntityTinyTNTPrimed primedTinyTNTEntity = new EntityTinyTNTPrimed( world, i + 0.5F, j + 0.5F,
+				k + 0.5F, (EntityLiving) null );
 		world.spawnEntityInWorld( primedTinyTNTEntity );
 		--dispensedItem.stackSize;
 		return dispensedItem;
