+++ src/main/java/appeng/items/misc/ItemCrystalSeed.java
@@ -63,7 +63,7 @@ public class ItemCrystalSeed extends AEBaseItem implements IGrowableCrystal
 	{
 		NBTTagCompound comp = Platform.openNbtData( is );
 		comp.setInteger( "progress", newDamage );
-		is.setItemDamage( (int) (is.getItemDamage() / LEVEL_OFFSET) * LEVEL_OFFSET );
+		is.setItemDamage( is.getItemDamage() / LEVEL_OFFSET * LEVEL_OFFSET );
 	}
 
 	public ItemCrystalSeed() {
